LookHubWeb/pyproject.toml
[project]
name = "lookhub"
version = "0.1.0"
description = ""
authors = [
    {name = "kibaidar1",email = "aidar1996@gmail.com"}
]
packages = [
  { include = "app" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.4.1,<9.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "asgiref (>=3.9.1,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
